break down devices file
