+++ packages/react-client/src/hooks/public.ts
@@ -1,6 +1,8 @@
 export { useConnect, useDisconnect } from "./useConnection";
 export { useReconnection } from "./useReconnection";
-export { useCamera, useMicrophone, useInitializeDevices } from "./devices";
+export { useCamera } from "./devices/useCamera";
+export { useMicrophone } from "./devices/useMicrophone";
+export { useInitializeDevices } from "./devices/useInitializeDevices";
 export { usePeers } from "./usePeers";
 export { useScreenShare } from "./useScreenShare";
 export { useStatus } from "./useStatus";
